requirements.txt
langchain
langchain-community
langchain-core
tavily-python
python-dotenv
streamlit
faiss-cpu
groq